requirements.txt
torch
numpy
PyYAML 